feat(pam): render /pam/requests/:id as a dialog over the tabs
The single-request detail was a full page, a sibling of the tabbed shell. Make it
a fourth child of the shell instead, hosting the detail in a dialog over it: the URL
is unchanged, so the five row links and the planned email deep link keep working and
still open in a new tab.

The route's component is now a host that opens the dialog and owns the navigation
that closing it needs - back when the caller came from inside the app, /pam when
they arrived cold. It renders the My requests tab behind the backdrop.
