+++ bitwarden_license/bit-web/src/app/pam/access-requests/access-request-route/access-request-route.component.html
@@ -1,159 +1,1 @@
-<app-header [title]="'pamAccessRequestTitle' | i18n"></app-header>
-
-@if (loading() && request() == null && !notFound() && !loadError()) {
-  <p bitTypography="body1" class="tw-text-muted">{{ "loading" | i18n }}</p>
-} @else if (request(); as r) {
-  <pam-request-summary
-    data-testid="request-detail-summary"
-    [cipher]="cipherFor(r.cipherId) ?? null"
-    [itemName]="cipherName() ?? ''"
-    [organizationName]="organizationName()"
-    [collectionName]="collectionName()"
-    [requesterName]="requesterDisplay() ?? ''"
-    [requesterEmail]="requesterEmail()"
-    [duration]="duration()"
-    [relativeStart]="start()"
-    [exactWindow]="exactWindowText()"
-    [reason]="reason()"
-  >
-    <pam-summary-field [label]="'pamColumnStatus' | i18n">
-      @if (badge()?.badgeState; as state) {
-        <app-pam-access-state-badge [state]="state" />
-      } @else if (badge()?.statusBadge; as b) {
-        <span bitBadge [variant]="b.variant">{{ b.labelKey | i18n }}</span>
-      }
-    </pam-summary-field>
-
-    <pam-summary-field [label]="'pamColumnSubmitted' | i18n">
-      {{ r.submittedAt | date: "medium" }}
-    </pam-summary-field>
-
-    @if (r.resolvedAt) {
-      <pam-summary-field [label]="'pamColumnResolved' | i18n">
-        {{ r.resolvedAt | date: "medium" }}
-      </pam-summary-field>
-    }
-  </pam-request-summary>
-
-  @if (r.producedLeaseId) {
-    <bit-section>
-      <bit-section-header>
-        <h2 bitTypography="h4">{{ "pamAccessRequestLeaseTitle" | i18n }}</h2>
-      </bit-section-header>
-      <dl class="tw-m-0 tw-flex tw-flex-col tw-gap-3">
-        <div class="tw-flex tw-gap-4">
-          <dt class="tw-w-40 tw-shrink-0 tw-font-bold">{{ "pamColumnStatus" | i18n }}</dt>
-          <dd class="tw-m-0 tw-flex-1">
-            @if (leaseActive()) {
-              <span bitBadge variant="success">{{ "pamStatusActivated" | i18n }}</span>
-            } @else if (badge()?.statusBadge; as b) {
-              <span bitBadge [variant]="b.variant">{{ b.labelKey | i18n }}</span>
-            }
-          </dd>
-        </div>
-        @if (showLeaseRemaining()) {
-          <div class="tw-flex tw-gap-4">
-            <dt class="tw-w-40 tw-shrink-0 tw-font-bold">{{ "pamColumnRemaining" | i18n }}</dt>
-            <dd class="tw-m-0 tw-flex-1">{{ r.leaseNotAfter | remainingTime: nowMs() }}</dd>
-          </div>
-        }
-      </dl>
-    </bit-section>
-  }
-
-  @if (decisions().length > 0) {
-    <bit-section>
-      <bit-section-header>
-        <h2 bitTypography="h4">{{ "pamAccessRequestDecisionLogTitle" | i18n }}</h2>
-      </bit-section-header>
-      <ul class="tw-m-0 tw-flex tw-list-none tw-flex-col tw-gap-4 tw-p-0">
-        @for (decision of decisions(); track $index) {
-          <li class="tw-flex tw-flex-col tw-gap-1">
-            <div class="tw-flex tw-items-center tw-gap-2">
-              <span
-                bitBadge
-                [variant]="
-                  decision.outcome === 'approved'
-                    ? 'success'
-                    : decision.outcome === 'denied'
-                      ? 'danger'
-                      : 'secondary'
-                "
-              >
-                {{ decision.labelKey | i18n }}
-              </span>
-              <span>
-                @if (decision.automatic) {
-                  {{ "pamResolverAccessRule" | i18n }}
-                } @else {
-                  {{ decision.who }}
-                }
-              </span>
-              <span class="tw-text-xs tw-text-muted">{{
-                decision.decidedAt | date: "medium"
-              }}</span>
-            </div>
-            @if (decision.comment) {
-              <div class="tw-text-xs tw-text-muted">{{ decision.comment }}</div>
-            }
-          </li>
-        }
-      </ul>
-    </bit-section>
-  }
-
-  <div class="tw-flex tw-flex-wrap tw-items-center tw-gap-2">
-    @if (canStart()) {
-      <button
-        type="button"
-        bitButton
-        buttonType="primary"
-        id="access-request_button_start"
-        [loading]="starting()"
-        [disabled]="starting()"
-        (click)="startAccess()"
-      >
-        {{ "pamStartLeaseButton" | i18n }}
-      </button>
-      <span class="tw-text-xs tw-text-muted">
-        {{ "pamActivateWithin" | i18n: (r.leaseNotAfter | remainingTime: nowMs()) }}
-      </span>
-    }
-    @if (canCancel()) {
-      <button
-        type="button"
-        bitButton
-        buttonType="secondary"
-        id="access-request_button_cancel"
-        [loading]="cancelling()"
-        [disabled]="cancelling()"
-        (click)="cancel()"
-      >
-        {{ "cancel" | i18n }}
-      </button>
-    }
-    @if (canEndLease()) {
-      <button
-        type="button"
-        bitButton
-        buttonType="secondary"
-        id="access-request_button_end-lease"
-        [loading]="ending()"
-        [disabled]="ending()"
-        (click)="endLease()"
-      >
-        {{ "pamEndLeaseButton" | i18n }}
-      </button>
-    }
-  </div>
-} @else if (notFound()) {
-  <bit-no-items>
-    <span slot="title" class="tw-mt-4 tw-block">{{ "pamAccessRequestNotFound" | i18n }}</span>
-    <span slot="description">{{ "pamAccessRequestNotFoundDescription" | i18n }}</span>
-    <a slot="button" bitButton buttonType="secondary" routerLink="/pam">
-      {{ "pamMyAccess" | i18n }}
-    </a>
-  </bit-no-items>
-} @else {
-  <p bitTypography="body1" class="tw-text-muted">{{ "pamAccessRequestLoadError" | i18n }}</p>
-}
+<pam-my-requests-tab />
